@@ -1,152 +1,91 @@
 {
   "name": "lx-music-desktop",
-  "version": "1.15.2",
+  "version": "2.5.0",
   "description": "一个免费的音乐查找助手",
-  "main": "./dist/electron/main.js",
+  "main": "./dist/main.js",
   "productName": "lx-music-desktop",
   "scripts": {
     "pack": "node build-config/pack.js && npm run pack:win:setup:x64",
     "pack:win": "node build-config/pack.js && npm run pack:win:setup:x64 && npm run pack:win:setup:x86 && npm run pack:win:setup:arm64 && npm run pack:win:setup:x86_64 && npm run pack:win:7z",
-    "pack:win:setup:x86_64": "cross-env TARGET=Setup ARCH=x86_64 electron-builder -w=nsis --x64 --ia32 -p never",
-    "pack:win:setup:x64": "cross-env TARGET=Setup ARCH=x64 electron-builder -w=nsis --x64 -p never",
-    "pack:win:setup:x86": "cross-env TARGET=Setup ARCH=x86 electron-builder -w=nsis --ia32 -p never",
-    "pack:win:setup:arm64": "cross-env TARGET=Setup ARCH=arm64 electron-builder -w=nsis --arm64 -p never",
+    "pack:win:setup:x86_64": "node build-config/build-pack.js target=win arch=x86_64 type=setup",
+    "pack:win:setup:x64": "node build-config/build-pack.js target=win arch=x64 type=setup",
+    "pack:win:setup:x86": "node build-config/build-pack.js target=win arch=x86 type=setup",
+    "pack:win:setup:arm64": "node build-config/build-pack.js target=win arch=arm64 type=setup",
     "pack:win:portable": "npm run pack:win:portable:x86_64 && npm run pack:win:portable:x64 && npm run pack:win:portable:x86",
-    "pack:win:portable:x86_64": "cross-env TARGET=portable ARCH=x86_64 electron-builder -w=portable --x64 --ia32 -p never",
-    "pack:win:portable:x64": "cross-env TARGET=portable ARCH=x64 electron-builder -w=portable --x64 -p never",
-    "pack:win:portable:x86": "cross-env TARGET=portable ARCH=x86 electron-builder -w=portable --ia32 -p never",
-    "pack:win:7z": "npm run pack:win:7z:x64 && npm run pack:win:7z:x86",
-    "pack:win:7z:x64": "cross-env TARGET=green ARCH=win_x64 electron-builder -w=7z --x64 -p never",
-    "pack:win:7z:x86": "cross-env TARGET=green ARCH=win_x86 electron-builder -w=7z --ia32 -p never",
-    "pack:win:7z:arm64": "cross-env TARGET=green ARCH=win_arm64 electron-builder -w=7z --arm64 -p never",
+    "pack:win:portable:x86_64": "node build-config/build-pack.js target=win arch=x86_64 type=portable",
+    "pack:win:portable:x64": "node build-config/build-pack.js target=win arch=x64 type=portable",
+    "pack:win:portable:x86": "node build-config/build-pack.js target=win arch=x86 type=portable",
+    "pack:win:7z": "npm run pack:win:7z:x64",
+    "pack:win:7z:x64": "node build-config/build-pack.js target=win arch=x64 type=green",
+    "pack:win:7z:arm64": "node build-config/build-pack.js target=win arch=arm64 type=green",
+    "pack:win7:7z:x64": "node build-config/build-pack.js target=win arch=x64 type=win7_green",
+    "pack:win7:7z:x86": "node build-config/build-pack.js target=win arch=x86 type=win7_green",
     "pack:linux": "node build-config/pack.js && npm run pack:linux:deb && npm run pack:linux:appImage && npm run pack:linux:rpm && npm run pack:linux:pacman",
-    "pack:linux:appImage": "cross-env ARCH=x64 electron-builder -l=AppImage -p never",
-    "pack:linux:deb": "npm run pack:linux:deb:x64 && npm run pack:linux:deb:x86 && npm run pack:linux:deb:arm64 && npm run pack:linux:deb:armv7l",
-    "pack:linux:deb:x64": "cross-env ARCH=x64 electron-builder -l=deb --x64 -p never",
-    "pack:linux:deb:x86": "cross-env ARCH=x86 electron-builder -l=deb --ia32 -p never",
-    "pack:linux:deb:arm64": "cross-env ARCH=arm64 electron-builder -l=deb --arm64 -p never",
-    "pack:linux:deb:armv7l": "cross-env ARCH=armv7l electron-builder -l=deb --armv7l -p never",
-    "pack:linux:rpm": "cross-env ARCH=x64 electron-builder -l=rpm --x64 -p never",
-    "pack:linux:pacman": "cross-env ARCH=x64 electron-builder -l=pacman --x64 -p never",
+    "pack:linux:appImage": "node build-config/build-pack.js target=linux arch=x64 type=appImage",
+    "pack:linux:deb": "npm run pack:linux:deb:amd64 && npm run pack:linux:deb:arm64 && npm run pack:linux:deb:armv7l",
+    "pack:linux:deb:amd64": "node build-config/build-pack.js target=linux arch=x64 type=deb",
+    "pack:linux:deb:arm64": "node build-config/build-pack.js target=linux arch=arm64 type=deb",
+    "pack:linux:deb:armv7l": "node build-config/build-pack.js target=linux arch=armv7l type=deb",
+    "pack:linux:rpm": "node build-config/build-pack.js target=linux arch=x64 type=rpm",
+    "pack:linux:pacman": "node build-config/build-pack.js target=linux arch=x64 type=pacman",
     "pack:mac": "node build-config/pack.js && npm run pack:mac:dmg && npm run pack:mac:dmg:arm64",
-    "pack:mac:dmg": "cross-env electron-builder -m=dmg -p never",
-    "pack:mac:dmg:arm64": "cross-env electron-builder -m=dmg --arm64 -p never",
-    "pack:dir": "node build-config/pack.js && electron-builder --dir",
+    "pack:mac:dmg": "node build-config/build-pack.js target=mac arch=x64 type=dmg",
+    "pack:mac:dmg:arm64": "node build-config/build-pack.js target=mac arch=arm64 type=dmg",
+    "pack:dir": "node build-config/pack.js && node build-config/build-pack.js target=dir",
     "publish": "node publish",
-    "publish:win:setup:x64:always": "cross-env TARGET=Setup ARCH=x64 electron-builder -w=nsis --x64 -p always",
-    "publish:win:setup:x64": "cross-env TARGET=Setup ARCH=x64 electron-builder -w=nsis --x64 -p always",
-    "publish:win:setup:x86": "cross-env TARGET=Setup ARCH=x86 electron-builder -w=nsis --ia32 -p onTagOrDraft",
-    "publish:win:setup:arm64": "cross-env TARGET=Setup ARCH=arm64 electron-builder -w=nsis --arm64 -p onTagOrDraft",
-    "publish:win:setup:x86_64": "cross-env TARGET=Setup ARCH=x86_64 electron-builder -w=nsis --x64 --ia32 -p onTagOrDraft",
+    "publish:win:setup:x64": "node build-config/build-pack.js target=win arch=x64 type=setup publish=always",
+    "publish:win:setup:x86": "node build-config/build-pack.js target=win arch=x86 type=setup publish=always",
+    "publish:win:setup:arm64": "node build-config/build-pack.js target=win arch=arm64 type=setup publish=always",
+    "publish:win:setup:x86_64": "node build-config/build-pack.js target=win arch=x86_64 type=setup publish=always",
     "publish:win:portable": "npm run publish:win:portable:x86_64 && npm run publish:win:portable:x64 && npm run publish:win:portable:x86",
-    "publish:win:portable:x86_64": "cross-env TARGET=portable ARCH=x86_64 electron-builder -w=portable --x64 --ia32 -p onTagOrDraft",
-    "publish:win:portable:x64": "cross-env TARGET=portable ARCH=x64 electron-builder -w=portable --x64 -p onTagOrDraft",
-    "publish:win:portable:x86": "cross-env TARGET=portable ARCH=x86 electron-builder -w=portable --ia32 -p onTagOrDraft",
-    "publish:win:7z:x64": "cross-env TARGET=green ARCH=win_x64 electron-builder -w=7z --x64 -p onTagOrDraft",
-    "publish:win:7z:x86": "cross-env TARGET=green ARCH=win_x86 electron-builder -w=7z --ia32 -p onTagOrDraft",
-    "publish:win:7z:arm64": "cross-env TARGET=green ARCH=win_arm64 electron-builder -w=7z --arm64 -p onTagOrDraft",
-    "publish:mac:dmg:always": "electron-builder -m=dmg -p always",
-    "publish:mac:dmg": "electron-builder -m=dmg -p onTagOrDraft",
-    "publish:mac:dmg:arm64": "electron-builder -m=dmg --arm64 -p onTagOrDraft",
-    "publish:linux:deb:x64:always": "cross-env ARCH=x64 electron-builder -l=deb --x64 -p always",
-    "publish:linux:deb:x64": "cross-env ARCH=x64 electron-builder -l=deb --x64 -p onTagOrDraft",
-    "publish:linux:deb:x86": "cross-env ARCH=x86 electron-builder -l=deb --ia32 -p onTagOrDraft",
-    "publish:linux:deb:arm64": "cross-env ARCH=arm64 electron-builder -l=deb --arm64 -p onTagOrDraft",
-    "publish:linux:deb:armv7l": "cross-env ARCH=armv7l electron-builder -l=deb --armv7l -p onTagOrDraft",
-    "publish:linux:appImage": "cross-env ARCH=x64 electron-builder -l=AppImage -p onTagOrDraft",
-    "publish:linux:rpm": "cross-env ARCH=x64 electron-builder -l=rpm --x64 -p onTagOrDraft",
-    "publish:linux:pacman": "cross-env ARCH=x64 electron-builder -l=pacman --x64 -p onTagOrDraft",
-    "dev": "node build-config/runner-dev.js",
-    "clean:electron": "rimraf dist/electron",
-    "clean": "rimraf dist && rimraf build",
-    "build:src": "node build-config/pack.js",
+    "publish:win:portable:x86_64": "node build-config/build-pack.js target=win arch=x86_64 type=portable publish=always",
+    "publish:win:portable:x64": "node build-config/build-pack.js target=win arch=x64 type=portable publish=always",
+    "publish:win:portable:x86": "node build-config/build-pack.js target=win arch=x86 type=portable publish=always",
+    "publish:win:7z:x64": "node build-config/build-pack.js target=win arch=x64 type=green publish=always",
+    "publish:win:7z:arm64": "node build-config/build-pack.js target=win arch=arm64 type=green publish=always",
+    "publish:win7:7z:x64": "node build-config/build-pack.js target=win arch=x64 type=win7_green publish=always",
+    "publish:win7:7z:x86": "node build-config/build-pack.js target=win arch=x86 type=win7_green publish=always",
+    "publish:mac:dmg": "node build-config/build-pack.js target=mac arch=x64 type=dmg publish=always",
+    "publish:mac:dmg:arm64": "node build-config/build-pack.js target=mac arch=arm64 type=dmg publish=always",
+    "publish:linux:deb:amd64": "node build-config/build-pack.js target=linux arch=x64 type=deb publish=always",
+    "publish:linux:deb:arm64": "node build-config/build-pack.js target=linux arch=arm64 type=deb publish=always",
+    "publish:linux:deb:armv7l": "node build-config/build-pack.js target=linux arch=armv7l type=deb publish=always",
+    "publish:linux:appImage": "node build-config/build-pack.js target=linux arch=x64 type=appImage publish=always",
+    "publish:linux:rpm": "node build-config/build-pack.js target=linux arch=x64 type=rpm publish=always",
+    "publish:linux:pacman": "node build-config/build-pack.js target=linux arch=x64 type=pacman publish=always",
+    "dev": "cross-env NODE_OPTIONS=--max-http-header-size=200000 node build-config/runner-dev.js",
+    "build:theme": "node src/common/theme/createThemes.js",
+    "build": "node build-config/pack.js",
     "build:main": "cross-env NODE_ENV=production webpack --config build-config/main/webpack.config.prod.js --progress",
     "build:renderer": "cross-env NODE_ENV=production webpack --config build-config/renderer/webpack.config.prod.js --progress",
     "build:renderer-lyric": "cross-env NODE_ENV=production webpack --config build-config/renderer-lyric/webpack.config.prod.js --progress",
-    "build": "npm run clean:electron && npm run build:main && npm run build:renderer && npm run build:renderer-lyric",
-    "lint": "eslint --ext .js,.vue -f ./node_modules/eslint-formatter-friendly src",
-    "lint:fix": "eslint --ext .js,.vue -f ./node_modules/eslint-formatter-friendly --fix src",
-    "dp": "cross-env ELECTRON_GET_USE_PROXY=true GLOBAL_AGENT_HTTPS_PROXY=http://localhost:1081 npm run pack",
-    "up": "cross-env ELECTRON_GET_USE_PROXY=true GLOBAL_AGENT_HTTPS_PROXY=http://localhost:1081 npm i"
+    "build:renderer-scripts": "cross-env NODE_ENV=production webpack --config build-config/renderer-scripts/webpack.config.prod.js --progress",
+    "lint": "eslint --ext .ts,.js,.vue -f node_modules/eslint-formatter-friendly src",
+    "lint:fix": "eslint --ext .ts,.js,.vue -f node_modules/eslint-formatter-friendly --fix src",
+    "postinstall": "electron-builder install-app-deps",
+    "dp": "cross-env ELECTRON_GET_USE_PROXY=true GLOBAL_AGENT_HTTPS_PROXY=http://127.0.0.1:2081 npm run pack",
+    "up": "cross-env ELECTRON_GET_USE_PROXY=true GLOBAL_AGENT_HTTPS_PROXY=http://127.0.0.1:2081 npm i"
   },
   "browserslist": [
-    "Electron 15.2.0"
+    "Electron 22.3.0"
   ],
   "engines": {
-    "node": ">= 14"
+    "node": ">= 16",
+    "npm": ">=8.5.2"
   },
-  "build": {
-    "appId": "cn.toside.music.desktop",
-    "directories": {
-      "buildResources": "./resources",
-      "output": "./build"
+  "macLanguagesInfoPlistStrings": {
+    "en": {
+      "CFBundleDisplayName": "LX Music",
+      "CFBundleName": "LX Music"
     },
-    "files": [
-      "dist/electron/**/*"
-    ],
-    "asar": {
-      "smartUnpack": false
+    "zh_CN": {
+      "CFBundleDisplayName": "LX Music",
+      "CFBundleName": "LX Music"
     },
-    "extraResources": [
-      "./licenses"
-    ],
-    "win": {
-      "icon": "./resources/icons/icon.ico",
-      "legalTrademarks": "lyswhut",
-      "artifactName": "${productName} v${version} ${env.ARCH} ${env.TARGET}.${ext}"
-    },
-    "mac": {
-      "icon": "./resources/icons/icon.icns",
-      "category": "public.app-category.music"
-    },
-    "linux": {
-      "maintainer": "lyswhut <lyswhut@qq.com>",
-      "artifactName": "${productName} v${version} ${env.ARCH}.${ext}",
-      "icon": "./resources/icons",
-      "category": "Utility;AudioVideo;Audio;Player;Music;",
-      "desktop": {
-        "Name[zh_CN]": "洛雪音乐助手"
-      }
-    },
-    "nsis": {
-      "oneClick": false,
-      "language": "2052",
-      "allowToChangeInstallationDirectory": true,
-      "differentialPackage": true,
-      "license": "./licenses/license.rtf",
-      "shortcutName": "LX Music"
-    },
-    "dmg": {
-      "window": {
-        "width": 600,
-        "height": 400
-      },
-      "contents": [
-        {
-          "x": 106,
-          "y": 252,
-          "name": "LX Music"
-        },
-        {
-          "x": 490,
-          "y": 252,
-          "type": "link",
-          "path": "/Applications"
-        }
-      ],
-      "title": "洛雪音乐助手 v${version}"
-    },
-    "appImage": {
-      "license": "./licenses/license_zh.txt",
-      "category": "Utility;AudioVideo;Audio;Player;Music;"
-    },
-    "publish": [
-      {
-        "provider": "github",
-        "owner": "lyswhut",
-        "repo": "lx-music-desktop"
-      }
-    ]
+    "zh_TW": {
+      "CFBundleDisplayName": "LX Music",
+      "CFBundleName": "LX Music"
+    }
   },
   "repository": {
     "type": "git",
@@ -167,88 +106,109 @@
   },
   "homepage": "https://github.com/lyswhut/lx-music-desktop#readme",
   "devDependencies": {
-    "@babel/core": "^7.16.0",
-    "@babel/plugin-proposal-class-properties": "^7.16.0",
+    "@babel/core": "^7.23.0",
+    "@babel/eslint-parser": "^7.22.15",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-modules-umd": "^7.16.0",
-    "@babel/plugin-transform-runtime": "^7.16.0",
-    "@babel/polyfill": "^7.12.1",
-    "@babel/preset-env": "^7.16.0",
-    "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.2.3",
-    "babel-preset-minify": "^0.5.1",
-    "browserslist": "^4.17.6",
-    "cfonts": "^2.10.0",
+    "@babel/plugin-transform-modules-umd": "^7.22.5",
+    "@babel/plugin-transform-runtime": "^7.22.15",
+    "@babel/preset-env": "^7.22.20",
+    "@babel/preset-typescript": "^7.23.0",
+    "@tsconfig/recommended": "^1.0.3",
+    "@types/better-sqlite3": "^7.6.5",
+    "@types/needle": "^3.2.1",
+    "@types/tunnel": "^0.0.4",
+    "@types/ws": "8.5.4",
+    "@volar/vue-language-plugin-pug": "^1.6.5",
+    "@vue/language-plugin-pug": "^1.8.15",
+    "babel-loader": "^9.1.3",
+    "browserslist": "^4.22.0",
     "chalk": "^4.1.2",
-    "changelog-parser": "^2.8.0",
-    "copy-webpack-plugin": "^9.0.1",
-    "core-js": "^3.19.1",
+    "changelog-parser": "^3.0.1",
+    "copy-webpack-plugin": "^11.0.0",
+    "core-js": "^3.32.2",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.5.1",
-    "css-minimizer-webpack-plugin": "^3.1.2",
-    "del": "^6.0.0",
-    "electron": "^13.4.0",
-    "electron-builder": "^22.11.7",
+    "css-loader": "^6.8.1",
+    "css-minimizer-webpack-plugin": "^5.0.1",
+    "del": "^6.1.1",
+    "electron": "^25.8.3",
+    "electron-builder": "^24.7.0",
     "electron-debug": "^3.2.0",
     "electron-devtools-installer": "^3.2.0",
-    "electron-to-chromium": "^1.3.891",
-    "eslint": "^7.32.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-formatter-friendly": "^7.0.0",
-    "eslint-plugin-html": "^6.2.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.1",
-    "eslint-plugin-standard": "^4.1.0",
-    "eslint-webpack-plugin": "^3.1.0",
-    "file-loader": "^6.2.0",
-    "friendly-errors-webpack-plugin": "^1.7.0",
-    "html-webpack-plugin": "^5.5.0",
-    "less": "^4.1.2",
-    "less-loader": "^10.2.0",
-    "less-plugin-clean-css": "^1.5.1",
-    "markdown-it": "^12.2.0",
-    "mini-css-extract-plugin": "^2.4.4",
+    "electron-to-chromium": "^1.4.532",
+    "electron-updater": "^6.1.5",
+    "eslint": "^8.50.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-config-standard-with-typescript": "^39.1.0",
+    "eslint-formatter-friendly": "github:lyswhut/eslint-friendly-formatter#2170d1320e2fad13615a9dcf229669f0bb473a53",
+    "eslint-plugin-html": "^7.1.0",
+    "eslint-plugin-vue": "^9.17.0",
+    "eslint-plugin-vue-pug": "^0.6.0",
+    "eslint-webpack-plugin": "^4.0.1",
+    "html-webpack-plugin": "^5.5.3",
+    "less": "^4.2.0",
+    "less-loader": "^11.1.3",
+    "mini-css-extract-plugin": "^2.7.6",
     "node-loader": "^2.0.0",
-    "postcss": "^8.3.11",
-    "postcss-loader": "^6.2.0",
+    "postcss": "^8.4.30",
+    "postcss-loader": "^7.3.3",
     "postcss-pxtorem": "^6.0.0",
     "pug": "^3.0.2",
-    "pug-loader": "^2.4.0",
     "pug-plain-loader": "^1.1.0",
-    "raw-loader": "^4.0.2",
-    "rimraf": "^3.0.2",
-    "spinnies": "^0.5.1",
-    "terser-webpack-plugin": "^5.2.5",
-    "url-loader": "^4.1.1",
-    "vue-loader": "^15.9.8",
-    "vue-template-compiler": "^2.6.14",
-    "webpack": "^5.62.1",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-hot-middleware": "^2.25.1",
-    "webpack-merge": "^5.8.0"
+    "rimraf": "^5.0.5",
+    "spinnies": "github:lyswhut/spinnies#233305c58694aa3b053e3ab9af9049993f918b9d",
+    "svg-sprite-loader": "^6.0.11",
+    "svg-transform-loader": "^2.0.13",
+    "svgo-loader": "^4.0.0",
+    "terser": "^5.20.0",
+    "terser-webpack-plugin": "^5.3.9",
+    "ts-loader": "^9.4.4",
+    "typescript": "^5.1.6",
+    "vue-eslint-parser": "^9.3.1",
+    "vue-loader": "^17.2.2",
+    "vue-template-compiler": "^2.7.14",
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1",
+    "webpack-hot-middleware": "github:lyswhut/webpack-hot-middleware#329c4375134b89d39da23a56a94db651247c74a1",
+    "webpack-merge": "^5.9.0"
   },
   "dependencies": {
-    "bufferutil": "^4.0.5",
+    "@simonwep/pickr": "^1.8.2",
+    "better-sqlite3": "^8.6.0",
+    "bufferutil": "^4.0.7",
+    "comlink": "~4.3.1",
     "crypto-js": "^4.1.1",
-    "electron-log": "^4.4.1",
-    "electron-store": "^8.0.1",
-    "electron-updater": "^4.3.9",
-    "font-list": "git+https://github.com/lyswhut/node-font-list.git#2ed3b4ee42e8a43373e8a30d87760c840725843e",
-    "http-terminator": "^3.0.3",
+    "electron-font-manager": "github:lyswhut/electron-font-manager#6d2f5ecf850c4fe34812b9394913680462ee0dae",
+    "electron-log": "^5.0.0-rc.1",
+    "electron-store": "^8.1.0",
+    "font-list": "^1.5.1",
     "iconv-lite": "^0.6.3",
-    "image-size": "^1.0.0",
-    "koa": "^2.13.4",
-    "long": "^5.1.0",
-    "needle": "^3.0.0",
-    "node-id3": "^0.2.3",
-    "request": "^2.88.2",
-    "socket.io": "^4.3.2",
-    "utf-8-validate": "^5.0.7",
-    "vue": "^2.6.14",
-    "vue-i18n": "^8.26.7",
-    "vue-router": "^3.5.3",
-    "vuex": "^3.6.2"
+    "image-size": "^1.0.2",
+    "jschardet": "^3.0.0",
+    "long": "^5.2.3",
+    "message2call": "^0.1.3",
+    "music-metadata": "^8.1.4",
+    "needle": "github:lyswhut/needle#93299ac841b7e9a9f82ca7279b88aaaeda404060",
+    "node-id3": "^0.2.6",
+    "sortablejs": "^1.15.0",
+    "tunnel": "^0.0.6",
+    "utf-8-validate": "^6.0.3",
+    "vue": "~3.2.47",
+    "vue-router": "^4.2.5",
+    "ws": "^8.14.2"
+  },
+  "overrides": {
+    "got": "^11",
+    "json5": "latest",
+    "minimatch": "latest",
+    "semver": "latest",
+    "svg-sprite-loader": {
+      "postcss": "latest"
+    },
+    "svg-baker": {
+      "postcss": "latest"
+    },
+    "http-cache-semantics": "latest"
   }
 }
