@@ -1,16 +1,23 @@
-const { contextBridge, ipcRenderer } = require('electron')
-const needle = require('needle')
-const { createCipheriv, publicEncrypt, constants, randomBytes, createHash } = require('crypto')
-const USER_API_RENDERER_EVENT_NAME = require('../rendererEvent/name')
+import { contextBridge, ipcRenderer } from 'electron'
+import needle from 'needle'
+import zlib from 'zlib'
+import { createCipheriv, publicEncrypt, constants, randomBytes, createHash } from 'crypto'
+import USER_API_RENDERER_EVENT_NAME from '../rendererEvent/name'
 
-const sendMessage = (action, status, data, message) => {
-  ipcRenderer.send(action, { status, data, message })
+for (const key of Object.keys(process.env)) {
+  if (/^(?:http_proxy|https_proxy|NO_PROXY)$/i.test(key)) delete process.env[key]
+}
+
+const sendMessage = (action, data, status, message) => {
+  ipcRenderer.send(action, { data, status, message })
 }
 
 let isInitedApi = false
+let isShowedUpdateAlert = false
 const EVENT_NAMES = {
   request: 'request',
   inited: 'inited',
+  updateAlert: 'updateAlert',
 }
 const eventNames = Object.values(EVENT_NAMES)
 const events = {
@@ -18,11 +25,11 @@ const events = {
 }
 const allSources = ['kw', 'kg', 'tx', 'wy', 'mg']
 const supportQualitys = {
-  kw: ['128k', '320k', 'flac'],
-  kg: ['128k', '320k', 'flac'],
-  tx: ['128k', '320k', 'flac'],
-  wy: ['128k', '320k', 'flac'],
-  mg: ['128k', '320k', 'flac'],
+  kw: ['128k', '320k', 'flac', 'flac24bit'],
+  kg: ['128k', '320k', 'flac', 'flac24bit'],
+  tx: ['128k', '320k', 'flac', 'flac24bit'],
+  wy: ['128k', '320k', 'flac', 'flac24bit'],
+  mg: ['128k', '320k', 'flac', 'flac24bit'],
 }
 const supportActions = {
   kw: ['musicUrl'],
@@ -35,7 +42,7 @@ const supportActions = {
 
 const handleRequest = (context, { requestKey, data }) => {
   // console.log(data)
-  if (!events.request) return sendMessage(USER_API_RENDERER_EVENT_NAME.response, false, { requestKey }, 'Request event is not defined')
+  if (!events.request) return sendMessage(USER_API_RENDERER_EVENT_NAME.response, { requestKey }, false, 'Request event is not defined')
   try {
     events.request.call(context, { source: data.source, action: data.action, info: data.info }).then(response => {
       let sendData = {
@@ -53,12 +60,12 @@ const handleRequest = (context, { requestKey, data }) => {
           }
           break
       }
-      sendMessage(USER_API_RENDERER_EVENT_NAME.response, true, sendData)
+      sendMessage(USER_API_RENDERER_EVENT_NAME.response, sendData, true)
     }).catch(err => {
-      sendMessage(USER_API_RENDERER_EVENT_NAME.response, false, { requestKey }, err.message)
+      sendMessage(USER_API_RENDERER_EVENT_NAME.response, { requestKey }, false, err.message)
     })
   } catch (err) {
-    sendMessage(USER_API_RENDERER_EVENT_NAME.response, false, { requestKey }, err.message)
+    sendMessage(USER_API_RENDERER_EVENT_NAME.response, { requestKey }, false, err.message)
   }
 }
 
@@ -70,17 +77,17 @@ const handleRequest = (context, { requestKey, data }) => {
  *                    status: true,
  *                    message: 'xxx',
  *                    sources: {
- *                         kw: ['128k', '320k', 'flac'],
- *                         kg: ['128k', '320k', 'flac'],
- *                         tx: ['128k', '320k', 'flac'],
- *                         wy: ['128k', '320k', 'flac'],
- *                         mg: ['128k', '320k', 'flac'],
+ *                         kw: ['128k', '320k', 'flac', 'flac24bit'],
+ *                         kg: ['128k', '320k', 'flac', 'flac24bit'],
+ *                         tx: ['128k', '320k', 'flac', 'flac24bit'],
+ *                         wy: ['128k', '320k', 'flac', 'flac24bit'],
+ *                         mg: ['128k', '320k', 'flac', 'flac24bit'],
  *                     }
  *                 }
  */
 const handleInit = (context, info) => {
   if (!info) {
-    sendMessage(USER_API_RENDERER_EVENT_NAME.init, false, null, 'Init failed')
+    sendMessage(USER_API_RENDERER_EVENT_NAME.init, null, false, 'Init failed')
     // sendMessage(USER_API_RENDERER_EVENT_NAME.init, false, null, typeof info.message === 'string' ? info.message.substring(0, 100) : '')
     return
   }
@@ -88,7 +95,7 @@ const handleInit = (context, info) => {
     sendMessage(USER_API_RENDERER_EVENT_NAME.openDevTools)
   }
   if (!info.status) {
-    sendMessage(USER_API_RENDERER_EVENT_NAME.init, false, null, 'Init failed')
+    sendMessage(USER_API_RENDERER_EVENT_NAME.init, null, false, 'Init failed')
     // sendMessage(USER_API_RENDERER_EVENT_NAME.init, false, null, typeof info.message === 'string' ? info.message.substring(0, 100) : '')
     return
   }
@@ -109,14 +116,26 @@ const handleInit = (context, info) => {
     }
   } catch (error) {
     console.log(error)
-    sendMessage(USER_API_RENDERER_EVENT_NAME.init, false, null, error.message)
+    sendMessage(USER_API_RENDERER_EVENT_NAME.init, null, false, error.message)
     return
   }
-  sendMessage(USER_API_RENDERER_EVENT_NAME.init, true, sourceInfo)
+  sendMessage(USER_API_RENDERER_EVENT_NAME.init, sourceInfo, true)
 
   ipcRenderer.on(USER_API_RENDERER_EVENT_NAME.request, (event, data) => {
     handleRequest(context, data)
   })
+}
+
+const handleShowUpdateAlert = (data, resolve, reject) => {
+  if (!data || typeof data != 'object') return reject(new Error('parameter format error.'))
+  if (!data.log || typeof data.log != 'string') return reject(new Error('log is required.'))
+  if (data.updateUrl && !/^https?:\/\/[^\s$.?#].[^\s]*$/.test(data.updateUrl) && data.updateUrl.length > 1024) delete data.updateUrl
+  if (data.log.length > 1024) data.log = data.log.substring(0, 1024) + '...'
+  sendMessage(USER_API_RENDERER_EVENT_NAME.showUpdateAlert, {
+    log: data.log,
+    updateUrl: data.updateUrl,
+  })
+  resolve()
 }
 
 contextBridge.exposeInMainWorld('lx', {
@@ -151,7 +170,7 @@ contextBridge.exposeInMainWorld('lx', {
         headers: resp.headers,
         bytes: resp.bytes,
         raw: resp.raw,
-        body: body,
+        body,
       }, body)
     }).request
 
@@ -165,12 +184,18 @@ contextBridge.exposeInMainWorld('lx', {
       if (!eventNames.includes(eventName)) return reject(new Error('The event is not supported: ' + eventName))
       switch (eventName) {
         case EVENT_NAMES.inited:
-          if (isInitedApi) return
+          if (isInitedApi) return reject(new Error('Script is inited'))
           isInitedApi = true
           handleInit(this, data)
+          resolve()
+          break
+        case EVENT_NAMES.updateAlert:
+          if (isShowedUpdateAlert) return reject(new Error('The update alert can only be called once.'))
+          isShowedUpdateAlert = true
+          handleShowUpdateAlert(data, resolve, reject)
           break
         default:
-          resolve(new Error('Unknown event name: ' + eventName))
+          reject(new Error('Unknown event name: ' + eventName))
       }
     })
   },
@@ -180,7 +205,9 @@ contextBridge.exposeInMainWorld('lx', {
       case EVENT_NAMES.request:
         events.request = handler
         break
+      default: return Promise.reject(new Error('The event is not supported: ' + eventName))
     }
+    return Promise.resolve()
   },
   utils: {
     crypto: {
@@ -190,7 +217,7 @@ contextBridge.exposeInMainWorld('lx', {
       },
       rsaEncrypt(buffer, key) {
         buffer = Buffer.concat([Buffer.alloc(128 - buffer.length), buffer])
-        return publicEncrypt({ key: key, padding: constants.RSA_NO_PADDING }, buffer)
+        return publicEncrypt({ key, padding: constants.RSA_NO_PADDING }, buffer)
       },
       randomBytes(size) {
         return randomBytes(size)
@@ -207,8 +234,26 @@ contextBridge.exposeInMainWorld('lx', {
         return Buffer.from(buf, 'binary').toString(format)
       },
     },
+    zlib: {
+      inflate(buf) {
+        return new Promise((resolve, reject) => {
+          zlib.inflate(buf, (err, data) => {
+            if (err) reject(new Error(err.message))
+            else resolve(data)
+          })
+        })
+      },
+      deflate(data) {
+        return new Promise((resolve, reject) => {
+          zlib.deflate(data, (err, buf) => {
+            if (err) reject(new Error(err.message))
+            else resolve(buf)
+          })
+        })
+      },
+    },
   },
-  version: '1.1.0',
+  version: '1.3.0',
   // removeEvent(eventName, handler) {
   //   if (!eventNames.includes(eventName)) return Promise.reject(new Error('The event is not supported: ' + eventName))
   //   let handlers
