@@ -3,10 +3,12 @@ const names = {
   request: '',
   response: '',
   openDevTools: '',
+  showUpdateAlert: '',
 }
 
 
 for (const key of Object.keys(names)) {
   names[key] = `userApi_${key}`
 }
-module.exports = names
+
+export default names
