@@ -1,20 +1,42 @@
 <template>
-  <Modal :show="visible" @close="handleCancel" @after-leave="afterLeave" :closeBtn="false">
-    <main :class="$style.main">{{message}}</main>
+  <Modal :show="visible" :close-btn="false" :teleport="teleport" @close="handleCancel" @after-leave="afterLeave">
+    <main class="scroll" :class="[$style.main, { 'select': selection }]">{{ message }}</main>
     <footer :class="$style.footer">
-      <Btn :class="$style.btn" v-if="showCancel" @click="handleCancel">{{cancelBtnText}}</Btn>
-      <Btn :class="$style.btn" @click="handleComfirm">{{confirmBtnText}}</Btn>
+      <Btn v-if="showCancel" :class="$style.btn" @click="handleCancel">{{ cancelBtnText }}</Btn>
+      <Btn :class="$style.btn" @click="handleComfirm">{{ confirmBtnText }}</Btn>
     </footer>
   </Modal>
 </template>
 
 <script>
-import Modal from '@renderer/components/material/Modal'
-import Btn from '@renderer/components/material/Btn'
+import Modal from '@renderer/components/material/Modal.vue'
+import Btn from '@renderer/components/base/Btn.vue'
+import { useI18n } from '@renderer/plugins/i18n'
+import { computed } from '@common/utils/vueTools'
 export default {
   components: {
     Modal,
     Btn,
+  },
+  props: {
+    afterLeave: {
+      type: Function,
+      default: () => {},
+    },
+  },
+  setup() {
+    const t = useI18n()
+
+    const defaultBtnTexts = computed(() => {
+      return {
+        confirm: t('confirm_button_text'),
+        cancel: t('cancel_button_text'),
+      }
+    })
+
+    return {
+      defaultBtnTexts,
+    }
   },
   data() {
     return {
@@ -23,29 +45,26 @@ export default {
       showCancel: false,
       cancelButtonText: '',
       confirmButtonText: '',
+      teleport: '#root',
+      selection: false,
     }
   },
   computed: {
     cancelBtnText() {
-      return this.cancelButtonText || this.$t('base.cancel_button_text')
+      return this.cancelButtonText || this.defaultBtnTexts.cancel
     },
     confirmBtnText() {
-      return this.confirmButtonText || this.$t('base.confirm_button_text')
+      return this.confirmButtonText || this.defaultBtnTexts.confirm
     },
   },
-  beforeDestroy() {
+  beforeUnmount() {
     const el = this.$el
     el.parentNode.removeChild(el)
   },
   methods: {
-    afterLeave(el, done) {
-      this.$destroy()
-    },
     handleCancel() {
-
     },
     handleComfirm() {
-
     },
   },
 }
@@ -56,9 +75,9 @@ export default {
 .main {
   flex: auto;
   min-height: 40px;
-  padding: 15px;
+  padding: 15px 15px 0;
   font-size: 14px;
-  max-width: 320px;
+  // max-width: 320px;
   min-width: 220px;
   line-height: 1.5;
   white-space: pre-line;
@@ -66,7 +85,7 @@ export default {
 
 .footer {
   flex: none;
-  padding: 0 15px 15px;
+  padding: 15px;
   display: flex;
   flex-flow: row nowrap;
   justify-content: flex-end;
