@@ -1,11 +1,22 @@
 <template>
   <transition name="tips-fade" @after-leave="afterLeave">
-    <div v-show="visible" :style="{ left: position.left + 'px' , top: position.top + 'px', transform: transform }" ref="dom_tips" :class="$style.tips">{{message}}</div>
+    <div
+      v-show="visible" ref="dom_tips" :style="{ left: position.left + 'px' , top: position.top + 'px', transform: transform }"
+      :class="$style.tips" role="presentation"
+    >
+      {{ message }}
+    </div>
   </transition>
 </template>
 
 <script>
 export default {
+  props: {
+    afterLeave: {
+      type: Function,
+      default: () => {},
+    },
+  },
   data() {
     return {
       visible: false,
@@ -27,14 +38,11 @@ export default {
       })
     },
   },
-  beforeDestroy() {
+  beforeUnmount() {
     const el = this.$el
     el.parentNode.removeChild(el)
   },
   methods: {
-    afterLeave(el, done) {
-      this.$destroy()
-    },
     handleGetOffsetXY(left, top) {
       const tipsWidth = this.$refs.dom_tips.clientWidth
       const tipsHeight = this.$refs.dom_tips.clientHeight
@@ -58,7 +66,7 @@ export default {
 </script>
 
 <style lang="less" module>
-@import '../../assets/styles/layout.less';
+@import '@renderer/assets/styles/layout.less';
 
 .tips {
   position: fixed;
@@ -66,16 +74,17 @@ export default {
   line-height: 1.2;
   word-wrap: break-word;
   padding: 4px 5px;
-  z-index: 999;
+  z-index: 10001;
   font-size: 12px;
   max-width: 80%;
-  color: @color-theme_2-font;
+  color: var(--color-font);
   border-radius: 3px;
-  background: @color-green-theme_2-background_1;
+  background: var(--color-content-background);
   overflow: hidden;
   pointer-events: none;
   // text-align: justify;
   box-shadow: 0 1px 8px rgba(0, 0, 0, 0.3);
+  white-space: pre;
 }
 
 :global(.tips-fade-enter-active), :global(.tips-fade-leave-active) {
@@ -84,15 +93,6 @@ export default {
 :global(.tips-fade-enter), :global(.tips-fade-leave-to) {
   opacity: 0;
 }
-
-each(@themes, {
-  :global(#container.@{value}) {
-    ~.tips {
-      color: ~'@{color-@{value}-theme_2-font}';
-      background: ~'@{color-@{value}-theme_2-background_1}';
-    }
-  }
-})
 
 
 </style>
