@@ -1,12 +1,24 @@
 const path = require('path')
 const ESLintPlugin = require('eslint-webpack-plugin')
 
+const isDev = process.env.NODE_ENV === 'development'
+
 module.exports = {
   target: 'electron-main',
   output: {
     filename: '[name].js',
-    libraryTarget: 'commonjs2',
-    path: path.join(__dirname, '../../dist/electron'),
+    library: {
+      type: 'commonjs2',
+    },
+    path: path.join(__dirname, '../../dist'),
+  },
+  externals: {
+    'font-list': 'font-list',
+    'better-sqlite3': 'better-sqlite3',
+    'electron-font-manager': 'electron-font-manager',
+    bufferutil: 'bufferutil',
+    'utf-8-validate': 'utf-8-validate',
+    'qrc_decode.node': isDev ? path.join(__dirname, '../../build/Release/qrc_decode.node') : path.join('../build/Release/qrc_decode.node'),
   },
   resolve: {
     alias: {
@@ -15,7 +27,7 @@ module.exports = {
       '@lyric': path.join(__dirname, '../../src/renderer-lyric'),
       '@common': path.join(__dirname, '../../src/common'),
     },
-    extensions: ['*', '.js', '.json', '.node'],
+    extensions: ['.tsx', '.ts', '.js', '.mjs', '.json', '.node'],
   },
   module: {
     rules: [
@@ -23,10 +35,12 @@ module.exports = {
         test: /\.node$/,
         use: 'node-loader',
       },
+      {
+        test: /\.tsx?$/,
+        use: 'ts-loader',
+        exclude: /node_modules/,
+      },
     ],
-  },
-  performance: {
-    maxEntrypointSize: 300000,
   },
   plugins: [
     new ESLintPlugin(),
