@@ -1,8 +1,7 @@
 const path = require('path')
-const VueLoaderPlugin = require('vue-loader/lib/plugin')
+const { VueLoaderPlugin } = require('vue-loader')
 const HTMLPlugin = require('html-webpack-plugin')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
-const CleanCSSPlugin = require('less-plugin-clean-css')
 const ESLintPlugin = require('eslint-webpack-plugin')
 
 const vueLoaderConfig = require('../vue-loader.config')
@@ -13,26 +12,39 @@ const isDev = process.env.NODE_ENV === 'development'
 module.exports = {
   target: 'electron-renderer',
   entry: {
-    'renderer-lyric': path.join(__dirname, '../../src/renderer-lyric/main.js'),
+    'renderer-lyric': path.join(__dirname, '../../src/renderer-lyric/main.ts'),
   },
   output: {
     filename: '[name].js',
-    libraryTarget: 'commonjs2',
-    path: path.join(__dirname, '../../dist/electron'),
-    publicPath: 'auto',
+    library: {
+      type: 'commonjs2',
+    },
+    path: path.join(__dirname, '../../dist'),
+    publicPath: '',
   },
   resolve: {
     alias: {
+      '@root': path.join(__dirname, '../../src'),
       '@main': path.join(__dirname, '../../src/main'),
       '@renderer': path.join(__dirname, '../../src/renderer'),
       '@lyric': path.join(__dirname, '../../src/renderer-lyric'),
       '@static': path.join(__dirname, '../../src/static'),
       '@common': path.join(__dirname, '../../src/common'),
     },
-    extensions: ['*', '.js', '.json', '.vue', '.node'],
+    extensions: ['.tsx', '.ts', '.js', '.json', '.node'],
   },
   module: {
     rules: [
+      {
+        test: /\.tsx?$/,
+        exclude: /node_modules/,
+        use: {
+          loader: 'ts-loader',
+          options: {
+            appendTsSuffixTo: [/\.vue$/],
+          },
+        },
+      },
       {
         test: /\.node$/,
         use: 'node-loader',
@@ -43,9 +55,8 @@ module.exports = {
         options: vueLoaderConfig,
       },
       {
-        test: /\.js$/,
-        loader: 'babel-loader',
-        exclude: /node_modules/,
+        test: /\.pug$/,
+        loader: 'pug-plain-loader',
       },
       {
         test: /\.css$/,
@@ -57,30 +68,12 @@ module.exports = {
           loader: 'less-loader',
           options: {
             sourceMap: true,
-            lessOptions: {
-              plugins: [
-                new CleanCSSPlugin({ advanced: true }),
-              ],
-            },
           },
         }),
       },
       {
-        test: /\.pug$/,
-        oneOf: [
-          // Use pug-plain-loader handle .vue file
-          {
-            resourceQuery: /vue/,
-            use: ['pug-plain-loader'],
-          },
-          // Use pug-loader handle .pug file
-          {
-            use: ['pug-loader'],
-          },
-        ],
-      },
-      {
         test: /\.(png|jpe?g|gif|svg)(\?.*)?$/,
+        exclude: path.join(__dirname, '../../src/renderer/assets/svgs'),
         type: 'asset',
         parser: {
           dataUrlCondition: {
@@ -90,6 +83,20 @@ module.exports = {
         generator: {
           filename: 'imgs/[name]-[contenthash:8][ext]',
         },
+      },
+      {
+        test: /\.svg$/,
+        include: path.join(__dirname, '../../src/renderer/assets/svgs'),
+        use: [
+          {
+            loader: 'svg-sprite-loader',
+            options: {
+              symbolId: 'icon-[name]',
+            },
+          },
+          'svg-transform-loader',
+          'svgo-loader',
+        ],
       },
       {
         test: /\.(mp4|webm|ogg|mp3|wav|flac|aac)(\?.*)?$/,
@@ -117,16 +124,12 @@ module.exports = {
       },
     ],
   },
-  performance: {
-    maxEntrypointSize: 300000,
-  },
   plugins: [
     new HTMLPlugin({
       filename: 'lyric.html',
-      template: path.join(__dirname, '../../src/renderer-lyric/index.pug'),
+      template: path.join(__dirname, '../../src/renderer-lyric/index.html'),
       isProd: process.env.NODE_ENV == 'production',
       browser: process.browser,
-      scriptLoading: 'blocking',
       __dirname,
     }),
     new VueLoaderPlugin(),
@@ -138,6 +141,7 @@ module.exports = {
     }),
     new ESLintPlugin({
       extensions: ['js', 'vue'],
+      formatter: require('eslint-formatter-friendly'),
     }),
   ],
 }
