@@ -6,24 +6,34 @@ const CopyWebpackPlugin = require('copy-webpack-plugin')
 const { merge } = require('webpack-merge')
 
 const baseConfig = require('./webpack.config.base')
+const buildConfig = require('../webpack-build-config')
 
-const { dependencies } = require('../../package.json')
+// const { dependencies } = require('../../package.json')
 
-let whiteListedModules = ['vue']
+// let whiteListedModules = ['vue', 'vue-router', 'vuex', 'vue-i18n']
 
 
 module.exports = merge(baseConfig, {
   mode: 'production',
   devtool: false,
   externals: [
-    ...Object.keys(dependencies || {}).filter(d => !whiteListedModules.includes(d)),
+    // ...Object.keys(dependencies || {}).filter(d => !whiteListedModules.includes(d)),
   ],
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        loader: 'babel-loader',
+        exclude: /node_modules/,
+      },
+    ],
+  },
   plugins: [
     new CopyWebpackPlugin({
       patterns: [
         {
           from: path.join(__dirname, '../../src/static'),
-          to: path.join(__dirname, '../../dist/electron/static'),
+          to: path.join(__dirname, '../../dist/static'),
         },
       ],
     }),
@@ -31,15 +41,25 @@ module.exports = merge(baseConfig, {
       'process.env': {
         NODE_ENV: '"production"',
       },
+      // ENVIRONMENT: 'process.env',
+      __VUE_OPTIONS_API__: 'true',
+      __VUE_PROD_DEVTOOLS__: 'false',
     }),
   ],
   optimization: {
+    minimize: buildConfig.minimize,
     minimizer: [
       new TerserPlugin(),
       new CssMinimizerPlugin(),
     ],
+    splitChunks: {
+      chunks: 'initial',
+      minChunks: 2,
+    },
   },
   performance: {
+    maxEntrypointSize: 1024 * 1024 * 10,
+    maxAssetSize: 1024 * 1024 * 20,
     hints: 'warning',
   },
   node: {
