@@ -1,22 +1,28 @@
 import tips from './Tips'
-import { debounce } from '../../utils'
+import { debounce } from '@common/utils'
 
 let instance
 let prevTips
 let prevX = 0
 let prevY = 0
+let isDraging = false
+
+const getTipText = el => {
+  return el.getAttribute('aria-label') && el.getAttribute('ignore-tip') == null ? el.getAttribute('aria-label') : null
+}
 
 const getTips = el =>
   el
-    ? el.getAttribute('tips')
-      ? el.getAttribute('tips')
+    ? getTipText(el)
+      ? getTipText(el)
       : el.parentNode === document.documentElement
         ? null
         : getTips(el.parentNode)
     : null
 
 const showTips = debounce(event => {
-  let msg = getTips(event.target)
+  if (isDraging) return
+  let msg = getTips(event.target)?.trim()
   if (!msg) return
   prevTips = msg
   instance = tips({
@@ -26,11 +32,12 @@ const showTips = debounce(event => {
       top: event.y + 12,
       left: event.x + 8,
     },
-  })
-  instance.$on('beforeClose', closeInstance => {
-    if (instance !== closeInstance) return
-    prevTips = null
-    instance = null
+  }, {
+    beforeClose(closeInstance) {
+      if (instance !== closeInstance) return
+      prevTips = null
+      instance = null
+    },
   })
 }, 400)
 
@@ -45,6 +52,7 @@ const setTips = tips => {
 }
 
 const updateTips = event => {
+  if (isDraging) return
   if (!instance) return showTips(event)
   setTimeout(() => {
     let msg = getTips(event.target)
@@ -54,14 +62,27 @@ const updateTips = event => {
   })
 }
 
-document.body.addEventListener('mousemove', event => {
-  if (event.x == prevX && event.y == prevY) return
-  prevX = event.x
-  prevY = event.y
-  hideTips()
-  showTips(event)
+setTimeout(() => {
+  document.body.addEventListener('mousemove', event => {
+    if ((event.x == prevX && event.y == prevY) || isDraging) return
+    prevX = event.x
+    prevY = event.y
+    hideTips()
+    showTips(event)
+  })
+
+  document.body.addEventListener('click', updateTips)
+
+  document.body.addEventListener('contextmenu', updateTips)
+
+  window.app_event.on('focus', () => {
+    hideTips()
+  })
+  window.app_event.on('dragStart', () => {
+    isDraging = true
+    hideTips()
+  })
+  window.app_event.on('dragEnd', () => {
+    isDraging = false
+  })
 })
-
-document.body.addEventListener('click', updateTips)
-
-document.body.addEventListener('contextmenu', updateTips)
